+++ packages/cubejs-api-gateway/UserError.js
@@ -1,0 +1,3 @@
+class UserError extends Error {}
+
+module.exports = UserError;