+++ packages/cubejs-query-orchestrator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.14](https://github.com/statsbotco/cubejs-client/compare/v0.9.13...v0.9.14) (2019-06-07)
+
+
+### Features
+
+* Add option to run in production without redis ([a7de417](https://github.com/statsbotco/cubejs-client/commit/a7de417)), closes [#110](https://github.com/statsbotco/cubejs-client/issues/110)
+
+
+
+
+
 ## [0.9.4](https://github.com/statsbotco/cubejs-client/compare/v0.9.3...v0.9.4) (2019-05-22)